requirements.txt
Django>=1.8,<1.9
django-betterforms==1.1.4
django-dynamic-preferences==0.8.2
django-extensions==1.7.4
django-modeladmin-reorder==0.2
django-registration==2.1.2
django-super-inlines==0.1.4
django-widget-tweaks==1.4.1
persisting-theory==0.2.1
six==1.10.0
